Move all data methods to one working file
As there was duplicated code before.

Stil lhaving problems with the breakpoint heights failing with a
conversion error.
