+++ _includes/county-map.html
@@ -19,9 +19,9 @@
   {% assign _viewbox_list = _viewbox | split: ' ' | to_f %}
   {% assign breakpoint_width = _viewbox_list[2] %}
   {% assign _height = _viewbox_list[3] %}
-  {% assign breakpoint_height = _viewbox_list[3] | times: 2.5 %}
+  {% assign breakpoint_height = _viewbox_list[3] | times: 2.5 | to_i %}
   {% capture is_wide %}
-    {% if breakpoint_width > breakpoint_height %}wide{% endif %}
+    wide
   {% endcapture %}
   {% assign is_wide = is_wide | strip %}
 {% endif %}
